refactor: extract downward helpers
